Rename omnibus-delegate.ts to vanilla-delegate.ts
