+++ src/services/direct/index.ts
@@ -5,4 +5,4 @@ export * from './common'
 export * from './health'
 export * from './fireblocks-provider'
 export * from './dfns-provider'
-export * from './omnibus-delegate'
+export * from './vanilla-delegate'
